+++ updater/omahahandler_test.go
@@ -6,11 +6,11 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/kinvolk/go-omaha/omaha"
+	"github.com/flatcar/go-omaha/omaha"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/kinvolk/nebraska/updater"
+	"github.com/flatcar/nebraska/updater"
 )
 
 const (